Move subclasses into their own folder
